tr-backend/ai-service/requirements.txt
fastapi
uvicorn
numpy
pillow
python-multipart
scikit-learn 
